agents/MCP Powered agents/travel-agent/requirements.txt
langchain==0.2.16
langchain-openai==0.1.23
langchain-core==0.2.39
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

